moved GroupByOrderByDimension to appropriate file
